mersiv-backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
motor==3.6.0
pymongo==4.9.2
python-socketio==5.11.4
python-multipart==0.0.18
pydantic==2.10.3
pydantic-settings==2.6.1
